Refactor: camera shared data init
